Completed holiday balance crud and fixed some bugs

+++ src/main/java/com/giantLink/RH/services/impl/HolidayBalanceServiceImpl.java
@@ -2,6 +2,7 @@ package com.giantLink.RH.services.impl;
 
 import com.giantLink.RH.entities.Employee;
 import com.giantLink.RH.entities.HolidayBalance;
+import com.giantLink.RH.exceptions.ResourceCantBeDeletedException;
 import com.giantLink.RH.exceptions.ResourceDuplicatedException;
 import com.giantLink.RH.exceptions.ResourceNotFoundException;
 import com.giantLink.RH.mappers.EmployeeMapper;
@@ -28,14 +29,14 @@ public class HolidayBalanceServiceImpl implements HolidayBalanceService {
 
     @Override
     public HolidayBalanceResponse add(HolidayBalanceRequest request) {
-//        Create the employee
+//        Create the holiday balance
         HolidayBalance holidayBalance = HolidayBalanceMapper.INSTANCE.requestToEntity(request);
 //        Save the holiday balance
         HolidayBalance holidayBalanceSaved = holidayBalanceRepository.save(holidayBalance);
 //        Link the holiday balance to the employee
         if (request.getEmployeeId() != null){
             Employee employee = employeeRepository.findById(request.getEmployeeId())
-                    .orElseThrow(() -> new ResourceNotFoundException("employee", "id", request.getId().toString()));
+                    .orElseThrow(() -> new ResourceNotFoundException("employee", "id", request.getEmployeeId().toString()));
             employee.setHolidayBalance(holidayBalanceSaved);
             employeeRepository.save(employee);
         }
@@ -65,8 +66,11 @@ public class HolidayBalanceServiceImpl implements HolidayBalanceService {
     @Override
     public void delete(Long id) {
 //        Check holiday balance exist
-        holidayBalanceRepository.findById(id)
+        HolidayBalance holidayBalance = holidayBalanceRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("holiday balance", "id", id.toString()));
+//        Check if it is linked to an employee
+        if (holidayBalance.getEmployee() != null)
+            throw new ResourceCantBeDeletedException("holiday balance", "it is linked to an employee");
 //        delete the holiday
         holidayBalanceRepository.deleteById(id);
     }
